requirements.txt
numpy
Pillow
pandas
reportlab

